renamet componenter og lagde ny victoryScren
